requirements.txt
fastapi
uvicorn
openai
tavily-python
python-dotenv
pydantic
sqlalchemy
